extract GroupedColumnButton to separate file
